+++ src/kspdg/__init__.py
@@ -1,4 +1,10 @@
-# Copyright (c) 2022, MASSACHUSETTS INSTITUTE OF TECHNOLOGY
+# Copyright (c) 2023, MASSACHUSETTS INSTITUTE OF TECHNOLOGY
 # Subject to FAR 52.227-11 – Patent Rights – Ownership by the Contractor (May 2014).
 # SPDX-License-Identifier: MIT
 
+# these imports make the individual environments accessible at the top-level
+# of the library and assign an environment version number
+# If the underlying environment changes, then the version number should be
+# incremented
+
+from kspdg.pe1.e1_envs import PE1_E1_I3_Env as PE1_E1_I3_V1